Database
Basic Database
Builds a database that stores the player's name and the total chips.
Used a test case in GameActivity.java that builds database with name John and total chips 100.
Still need to implement a function that gets the players name when they open the app, updates the chip after each round, and finally a interface that stores the top 10 names with the highest chips.